+++ lib/polarToCartesian.hpp
@@ -4,13 +4,13 @@
 namespace ngPlasmid
 {
     Point polarToCartesian(
-        double centerX,
-        double centerY,
-        double radius,
-        double angleInDegrees
+        long double centerX,
+        long double centerY,
+        long double radius,
+        long double angleInDegrees
     ) {
         Point res;
-        double angleInRadians = (angleInDegrees - 90) * M_PI / 180.0;
+        long double angleInRadians = (angleInDegrees - 90) * M_PI / 180.0;
         res.x = centerX + (radius * std::cos(angleInRadians));
         res.y = centerY + (radius * std::sin(angleInRadians));
         return res;
